💄 Add Doodle images to collectionView
