+++ PhotosApp/PhotosApp/Base.lproj/Main.storyboard
@@ -71,20 +71,34 @@
                         <rect key="frame" x="0.0" y="0.0" width="390" height="844"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-                        <collectionViewFlowLayout key="collectionViewLayout" automaticEstimatedItemSize="YES" minimumLineSpacing="10" minimumInteritemSpacing="10" id="vhz-KM-bYe">
-                            <size key="itemSize" width="128" height="128"/>
+                        <collectionViewFlowLayout key="collectionViewLayout" minimumLineSpacing="10" minimumInteritemSpacing="10" id="vhz-KM-bYe">
+                            <size key="itemSize" width="110" height="50"/>
                             <size key="headerReferenceSize" width="0.0" height="0.0"/>
                             <size key="footerReferenceSize" width="0.0" height="0.0"/>
                             <inset key="sectionInset" minX="0.0" minY="0.0" maxX="0.0" maxY="0.0"/>
                         </collectionViewFlowLayout>
                         <cells>
-                            <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" id="Iu0-Rb-v9k">
-                                <rect key="frame" x="0.0" y="0.0" width="128" height="128"/>
+                            <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" reuseIdentifier="DoodleViewCell" id="Iu0-Rb-v9k" customClass="DoodleViewCell" customModule="PhotosApp" customModuleProvider="target">
+                                <rect key="frame" x="0.0" y="0.0" width="110" height="50"/>
                                 <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
                                 <collectionViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" insetsLayoutMarginsFromSafeArea="NO" id="CkP-bB-cBf">
-                                    <rect key="frame" x="0.0" y="0.0" width="128" height="128"/>
+                                    <rect key="frame" x="0.0" y="0.0" width="110" height="50"/>
                                     <autoresizingMask key="autoresizingMask"/>
+                                    <subviews>
+                                        <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="zmw-NK-xYn">
+                                            <rect key="frame" x="0.0" y="0.0" width="110" height="50"/>
+                                        </imageView>
+                                    </subviews>
+                                    <constraints>
+                                        <constraint firstAttribute="bottom" secondItem="zmw-NK-xYn" secondAttribute="bottom" id="JN6-nI-5IA"/>
+                                        <constraint firstItem="zmw-NK-xYn" firstAttribute="top" secondItem="CkP-bB-cBf" secondAttribute="top" id="VHW-lw-N25"/>
+                                        <constraint firstItem="zmw-NK-xYn" firstAttribute="leading" secondItem="CkP-bB-cBf" secondAttribute="leading" id="YXT-Bw-XQz"/>
+                                        <constraint firstAttribute="trailing" secondItem="zmw-NK-xYn" secondAttribute="trailing" id="oT7-pL-yF6"/>
+                                    </constraints>
                                 </collectionViewCellContentView>
+                                <connections>
+                                    <outlet property="imageView" destination="zmw-NK-xYn" id="C26-zz-3rU"/>
+                                </connections>
                             </collectionViewCell>
                         </cells>
                         <connections>
