+++ PhotosApp/PhotosApp/DoodleViewCell.swift
@@ -1,0 +1,26 @@
+//
+//  DoodleViewCell.swift
+//  PhotosApp
+//
+//  Created by Song on 2021/03/25.
+//
+
+import UIKit
+
+class DoodleViewCell: UICollectionViewCell {
+    @IBOutlet weak var imageView: UIImageView!
+    
+    //override init(frame: CGRect) {
+    //    super.init(frame: frame)
+    //    commonInit()
+    //}
+    //
+    //required init?(coder: NSCoder) {
+    //    super.init(coder: coder)
+    //    commonInit()
+    //}
+    //
+    //func commonInit() {
+    //    imageView.contentMode = .scaleAspectFill
+    //}
+}
